+++ CHANGELOG.md
@@ -1,5 +1,14 @@
-## 4.4.2
+## 4.7.1
  - Improved action call-out of log warning when this plugin cannot enforce timeouts [#93](https://github.com/logstash-plugins/logstash-filter-kv/pull/93)
+
+## 4.7.0
+ - Allow attaching multiple tags on failure. The `tag_on_failure` option now also supports an array of strings [#100](https://github.com/logstash-plugins/logstash-filter-kv/issues/100). Fixes [#92](https://github.com/logstash-plugins/logstash-filter-kv/issues/92)
+
+## 4.6.0
+ - Added `allow_empty_values` option [#72](https://github.com/logstash-plugins/logstash-filter-kv/pull/72)
+
+## 4.5.0
+ - Feat: check that target is set in ECS mode [#96](https://github.com/logstash-plugins/logstash-filter-kv/pull/96)
 
 ## 4.4.1
  - Fixed issue where a `field_split_pattern` containing a literal backslash failed to match correctly [#87](https://github.com/logstash-plugins/logstash-filter-kv/issues/87)